refactor: TelephonyManagerHelper is introduced
