+++ app/src/main/java/com/hasib/callerid/di/SystemModule.kt
@@ -1,0 +1,21 @@
+package com.hasib.callerid.di
+
+import android.content.Context
+import android.telephony.TelephonyManager
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object SystemModule {
+
+    @Provides
+    @Singleton
+    fun getTelephonyManager(@ApplicationContext context: Context): TelephonyManager {
+        return context.getSystemService(Context.TELEPHONY_SERVICE) as TelephonyManager
+    }
+}
