+++ gradle/libs.versions.toml
@@ -1,6 +1,6 @@
 [versions]
 agp = "8.7.0"
-kotlin = "1.9.24"
+kotlin = "2.0.0"
 coreKtx = "1.15.0"
 junit = "4.13.2"
 junitVersion = "1.2.1"
